requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.6.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
Fabric3==1.14.post1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
paramiko==2.7.1
Pillow==6.2.2
psycopg2==2.8.4
pycparser==2.20
PyNaCl==1.3.0
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7.1
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
